backend_backup_upgrade_20250811_161943/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0
openai>=1.0.0
anthropic>=0.34.0
gunicorn>=21.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
stripe>=7.0.0
python-dotenv>=1.0.0
sendgrid>=6.10.0
psycopg2-binary
pyotp>=2.8.0
qrcode>=7.4.0
pillow>=10.0.0
geoip2>=4.7.0
user-agents>=2.2.0
passlib>=1.7.4
schedule>=1.2.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.5
flask-migrate>=4.0.5
firebase-admin>=6.2.0
requests>=2.31.0
werkzeug>=2.3.7
numpy>=1.24.0
scikit-learn>=1.3.0
watchdog>=3.0.0
